+++ src/pages/Login.tsx
@@ -1,9 +1,10 @@
-import React, { useState } from 'react';
+import React, { useRef, useState } from 'react';
 import { useHistory } from "react-router-dom";
 import AWS from 'aws-sdk/global';
 import { AuthenticationDetails, CognitoUserPool, CognitoUser } from 'amazon-cognito-identity-js';
 import Alignment from '@components/Alignment';
 import Form from '@components/Form';
+import Icon from '@components/Icon';
 import Input from '@components/Input';
 import Spinner from '@components/Spinner';
 import config from '@env';
@@ -14,10 +15,18 @@ type LoginProps = {
 };
 
 const Login: React.FC<LoginProps> = ({ setIsAuthorized }: LoginProps): JSX.Element => {
+  const [didAuthFail, setDidAuthFail] = useState(false);
   const [inputs, setInputs] = useState({ email: '', password: '' });
   const [isLoading, setIsLoading] = useState(false);
-  const [didAuthFail, setDidAuthFail] = useState(false);
+  const [isNewPasswordFocused, setIsNewPasswordFocused] = useState(false);
   const [isSignup, setIsSignup] = useState(false);
+  const [hasLowercaseCharacter, setHasLowercaseCharacter] = useState(false);
+  const [hasNumber, setHasNumber] = useState(false);
+  const [hasSpecialCharacter, setHasSpecialCharacter] = useState(false);
+  const [hasUppercaseCharacter, setHasUppercaseCharacter] = useState(false);
+  const [is8CharactersLong, setIs8CharactersLong] = useState(false);
+
+  const passwordRef = useRef<HTMLInputElement>(null);
 
   const {
     clientID,
@@ -81,8 +90,30 @@ const Login: React.FC<LoginProps> = ({ setIsAuthorized }: LoginProps): JSX.Eleme
       setInputs(prevState => ({ ...prevState, [kind]: event.target.value }));
     };
 
-  const handleButtonClick = (label: string): void =>
-    label === 'signup' ? setIsSignup(true) : setIsSignup(false);
+  const handleButtonClick = (flag: string): void =>
+    flag === 'signup' ? setIsSignup(true) : setIsSignup(false);
+
+  const handleFocus = (): void => setIsNewPasswordFocused(!isNewPasswordFocused);
+
+  const validatePassword = (): void => {
+    const refValue = passwordRef.current!.value;
+
+    if ((refValue.toUpperCase() !== refValue) !== hasLowercaseCharacter) {
+      setHasLowercaseCharacter(!hasLowercaseCharacter);
+    }
+    if (((/\d/).test(refValue)) !== hasNumber) {
+      setHasNumber(!hasNumber);
+    }
+    if (((/[\s~`!@#$%^&*+=\-[\]\\';,/{}|\\":<>?()._]/g).test(refValue)) !== hasSpecialCharacter) {
+      setHasSpecialCharacter(!hasSpecialCharacter);
+    }
+    if ((refValue.toLowerCase() !== refValue) !== hasUppercaseCharacter) {
+      setHasUppercaseCharacter(!hasUppercaseCharacter);
+    }
+    if ((refValue.length >= 8) !== is8CharactersLong) {
+      setIs8CharactersLong(!is8CharactersLong);
+    }
+  };
 
   return (
     <Alignment vertical horizontal>
@@ -91,18 +122,41 @@ const Login: React.FC<LoginProps> = ({ setIsAuthorized }: LoginProps): JSX.Eleme
           <h1>TIMEWARD</h1>
           <hr />
           <h2>DAY MANAGER & TIME TRACKER</h2>
-          {didAuthFail &&
-            <p>Incorrect username or password.</p>
+          {isNewPasswordFocused && isSignup &&
+            <div className='password-requirements'>
+              <p><Icon type={is8CharactersLong ? 'checkmark' : 'x'} />
+                Password must be at least 8 characters long.
+              </p>
+              <p><Icon type={hasNumber ? 'checkmark' : 'x'} />
+                Password must have at least 1 number.
+              </p>
+              <p><Icon type={hasSpecialCharacter ? 'checkmark' : 'x'} />
+                Password must have at least 1 special character.
+              </p>
+              <p><Icon type={hasLowercaseCharacter ? 'checkmark' : 'x'} />
+                Password must have at least 1 lowercase character.
+              </p>
+              <p><Icon type={hasUppercaseCharacter ? 'checkmark' : 'x'} />
+                Password must have at least 1 uppercase character.
+              </p>
+            </div>
           }
+          {didAuthFail && <p className='failure'>Incorrect username or password.</p>}
           <Form isSignup={isSignup} handleButtonClick={handleButtonClick} onSubmit={handleAuthFlow}>
             <Input
               onChange={handleInputChange('email')}
               placeholder={`${isSignup ? 'New ' : ''}Email`}
             />
             <Input
-              onChange={handleInputChange('password')}
+              onBlur={handleFocus}
+              onChange={(): void => {
+                handleInputChange('password');
+                if (isSignup) validatePassword();
+              }}
+              onFocus={handleFocus}
               placeholder={`${isSignup ? 'New ' : ''}Password`}
               password
+              ref={passwordRef}
             />
           </Form>
         </StyledLogin>